# packages/contracts/src/openapi.yaml
openapi: 3.1.0
info:
  title: Xenitex Appliance API
  version: 1.0.0
  description: >
    v1 of the on-premise appliance API. This document is the single source of
    truth (P1-01) — the TypeScript client and types in this package are
    generated from it, never hand-written (P1-02).

    Conventions (see docs/adr/0006-api-conventions.md, accepted at GATE 1):
    opaque cursor pagination on every collection (`?cursor=&limit=`, response
    `{ items, nextCursor }`); RFC 9457 Problem Details on every non-2xx
    response; `Idempotency-Key` required on every creation POST;
    `ETag`/`If-Match` required on every mutation of a mutable resource; URI
    path versioning (`/v1/...`); long-running operations return `202` with a
    `Location` header pointing at a pollable resource, never a blocking
    request.

    JSON casing: camelCase, matching packages/domain. Any field whose value
    originates from a scanned target (SEC-17/AIP-02 — banners, headers,
    certificate subjects, hostnames, identity-key values, OS/service/product
    inference) carries an `Untrusted` suffix in its property name. This is
    the same greppable-naming convention as the `_untrusted` column suffix in
    docs/adr/0007-schema-conventions.md, extended to the API boundary so a
    frontend consumer can never mistake such a value for first-party data —
    it must be escaped on output, never rendered as HTML, never interpreted.
servers:
  - url: /v1
...
security:
  - sessionCookie: []
paths:
...
  /observations/{observationId}:
    parameters:
      - name: observationId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Observations]
      operationId: getObservation
      summary: Evidence detail (MOD-21) — the matched banner, response excerpt, or port state, linked to its raw artifact
      responses:
        '200':
          {
            description: 'OK.',
            content: { application/json: { schema: { $ref: '#/components/schemas/Observation' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /raw-artifacts/{rawArtifactId}/download:
    parameters:
      - name: rawArtifactId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Observations]
      operationId: downloadRawArtifact
      summary: Download the immutable raw scanner artifact an observation is backed by (DATA-01)
      responses:
        '200':
          description: Raw artifact bytes.
          content: { application/octet-stream: { schema: { type: string, format: binary } } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Issues
...
  /intel/imports:
    get:
      tags: [Intelligence]
      operationId: listIntelImports
      summary: UI-106 — sync history
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/IntelImport' } }
                  nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
  /intel/sync:
    post:
      tags: [Intelligence]
      operationId: createIntelSync
      summary: 'FEED-11 — "Update now": a bounded incremental NVD+EPSS sync, staged then applied atomically (FEED-25)'
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                modifiedSinceDays: { type: integer, minimum: 1, maximum: 120, default: 7 }
      responses:
        '202':
          {
            description: 'Accepted — the sync runs asynchronously in apps/worker.',
            content: { application/json: { schema: { $ref: '#/components/schemas/IntelImport' } } },
          }
        '409':
          description: Online updates are permanently disabled for this deployment (FEED-23).
          content:
            {
              application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } },
            }
  /intel/settings:
    get:
      tags: [Intelligence]
      operationId: getIntelSettings
      summary: FEED-23 — the online-update disable switch
      responses:
        '200':
          {
            description: 'OK.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/IntelSettings' } } },
          }
    patch:
      tags: [Intelligence]
      operationId: updateIntelSettings
      summary: FEED-23
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [onlineUpdatesDisabled]
              properties: { onlineUpdatesDisabled: { type: boolean } }
      responses:
        '200':
          {
            description: 'OK.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/IntelSettings' } } },
          }
        '409': { $ref: '#/components/responses/StaleResource' }

  # ------------------------------------------------------------- SystemHealth
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xenitex_session
      description: >
        Opaque server-side session (SEC-08). State-changing requests also
        require the double-submit CSRF token in the `X-CSRF-Token` header,
        matched against the session's csrfTokenHash server-side.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`. Never a raw offset (ADR 0006).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current ETag of the resource. A stale value returns 409 (ADR 0006).
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key. A retried request with the same key returns the original response rather than creating a second resource (ADR 0006).
      schema: { type: string }
...
  headers:
...
    CorrelationId:
      description: >-
        OPS-01 — the correlation identifier for this request, propagated from
        the HTTP request through any queue job and scanner invocation it
        triggers. Echoes an inbound `X-Correlation-Id` if the client sent
        one, otherwise server-generated. Surfaced on error responses so a UI
        can quote it back in a support conversation (see ErrorState).
      schema: { type: string }
  responses:
...
    NotFound:
      description: No such resource.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    StaleResource:
      description: 'code: concurrency.stale_resource — the supplied If-Match no longer matches (ADR 0006).'
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable, machine-readable, namespaced (e.g. `scope.exclusion_violation`). Clients branch on this, never on title/detail (ADR 0006).
...
    Observation:
      type: object
      required:
        [
          id,
          scanRunId,
          scannerAdapterKey,
          scannerAdapterVersion,
          rawArtifactId,
          targetAddress,
          extractedAttributes,
          untrustedEvidence,
          observedAt,
        ]
      properties:
        id: { type: string }
        scanRunId: { type: string }
        scannerAdapterKey: { type: string }
        scannerAdapterVersion: { type: string }
        rawArtifactId: { type: string }
        targetAddress: { type: string }
        targetPort: { oneOf: [{ type: integer }, { type: 'null' }] }
        targetProtocol: { oneOf: [{ type: string }, { type: 'null' }] }
        resolvedAssetId: { oneOf: [{ type: string }, { type: 'null' }] }
        extractedAttributes: { type: object, additionalProperties: true }
        untrustedEvidence:
          type: object
          additionalProperties: { type: string }
          description: SEC-17 — every value here originated from the scanned target. Escape on output, never render as HTML, never interpret.
        observedAt: { type: string, format: date-time }
...
    IntelImport:
      type: object
      required: [id, sourceName, sourceVersion, status, importedAt]
      properties:
        id: { type: string }
        sourceName: { type: string }
        sourceVersion: { type: string }
        status: { type: string, enum: [validating, applied, failed, superseded] }
        fetched: { oneOf: [{ type: integer }, { type: 'null' }] }
        added: { oneOf: [{ type: integer }, { type: 'null' }] }
        modified: { oneOf: [{ type: integer }, { type: 'null' }] }
        discarded: { oneOf: [{ type: integer }, { type: 'null' }] }
        failureReason: { oneOf: [{ type: string }, { type: 'null' }] }
        importedAt: { type: string, format: date-time }
        importedBy: { oneOf: [{ type: string }, { type: 'null' }] }
        supersededById: { oneOf: [{ type: string }, { type: 'null' }] }
...
    IntelSettings:
      type: object
      required: [onlineUpdatesDisabled]
      properties:
        onlineUpdatesDisabled: { type: boolean }
...